requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==21.11b1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
dataclasses-json==0.5.6
decorate-url==1.1
decorator==4.4.2
DistributedLock==1.2
Django==3.2.9
django-annoying==0.10.6
django-appconf==1.0.5
django-compressor==3.1
django-cors-headers==3.10.0
django-crispy-forms==1.14.0
django-decorator-include==3.0
django-hash-field==0.1
django-oauth-tokens==0.6.3
django-sass-processor==1.1
django-sslserver==0.22
django-taggit==2.1.0
django-unicorn==0.39.1
djangorestframework==3.12.4
facebook-sdk==3.1.0
firebase-admin==5.1.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-apis-oauth-django==1.0.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
httplib2==0.20.2
idna==3.3
libsass==0.21.0
marshmallow==3.14.1
marshmallow-enum==1.5.1
msgpack==1.0.3
mypy-extensions==0.4.3
oauthlib==3.1.1
orjson==3.6.5
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
proto-plus==1.19.9
protobuf==3.19.3
psutil==5.9.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-facebook-api==0.12.4
python-memcached==1.59
pytz==2021.3
rcssmin==1.1.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.2.0
rsa==4.8
shortuuid==1.0.8
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==1.2.3
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
